requirements.txt
openai-whisper
ffmpeg-python
fastapi
uvicorn
python-multipart
